Applying typecast safely on Color class.

+++ src/com/mojang/ld22/gfx/Color.java
@@ -8,10 +8,10 @@ public class Color {
 	}
 
 	public static int get(int d) {
-		int red = d / 100 % 10;
-		int green = d / 10 % 10;
-		int blue = d % 10;
-		int colorMixture = red * 36 + green * 6 + blue;
+		int red = (int) d / 100 % 10;
+		int green = (int) d / 10 % 10;
+		int blue = (int) d % 10;
+		int colorMixture = ((int) red * 36 + green * 6 + blue);
 		
 		if (d < 0) {
 			return 255;
